Code cleaned
The code has been cleaned a little bit and there is a new rule that says: "if a player falls into an occupied square, the player that was first in the square is moved to a random position.".
It remains to make a case when a player only passes over another because, when that happens, one player is deleted and we don't want it.